Fix spell checker unit test fail.
The spell checker was adding an entry into the database for
"correct spelling" -> "correct spelling" which I couldn't find
any use for and which was causing the test to fail.  Removed.

Also removed three files from version control which are
generated by the testrunner program rather than being source
files.


git-svn-id: https://svn.code.sf.net/p/chessx/code/trunk@502 1de841de-8530-0410-b3ab-d372fa6ee214