Fix atSquare() calculation.
A subtle artifact of the way we calculated the square the mouse
is currently over meant we missed an opportunity to send an
"InvalidSquare" before the mouse left the application window.
This had negative consequences for the guess-the-move feature,
where the guessed move would remain displayed on the board
even when the mouse had moved away.  Fixed with this change.


git-svn-id: https://svn.code.sf.net/p/chessx/code/trunk@516 1de841de-8530-0410-b3ab-d372fa6ee214